# Vectorised AggregatedParameter.