Here's a benchmark from ASV

```
    before     after       ratio
  [919dd7e7] [5e2e9cfc]
-    1.32μs   802.27ns      0.61  benchmarks.AggregatedParameter.time_calc_values(10)
-    8.50μs     1.18μs      0.14  benchmarks.AggregatedParameter.time_calc_values(100)
-   40.52μs     2.68μs      0.07  benchmarks.AggregatedParameter.time_calc_values(500)
```