@snorfalorpagus 2017-03-15
The appveyor failures are usual conda rubbish. Need to test on Windows before merge.

This looks good to me. I was worried that we'd have two implementations of the algorithm (one for value one for all values) but I see you've just removed value. I think we should still implement value, but have it raise an error (something bespoke) to avoid tripping up.