@jetuk 2017-03-15
> I think we should still implement value, but have it raise an error (something bespoke) to avoid tripping up.

We discussed that the default `value` method will raise an error already.